Move utility traits to a `utils` module
