+++ src/entry.rs
@@ -1,6 +1,7 @@
 use xml::Element;
 
-use {Author, Category, Contributor, ElementUtils, Link, NS, Person, Source, ViaXml};
+use {Author, Category, Contributor, Link, NS, Person, Source};
+use utils::{ElementUtils, ViaXml};
 
 
 /// [The Atom Syndication Format § The "atom:entry" Element]
