+++ src/generator.rs
@@ -2,7 +2,8 @@ use std::convert::AsRef;
 
 use xml::Element;
 
-use {ElementUtils, NS, ViaXml};
+use NS;
+use utils::{ElementUtils, ViaXml};
 
 
 /// [The Atom Syndication Format § The "atom:generator" Element]
