+++ src/lib.rs
@@ -72,8 +72,7 @@ mod generator;
 mod link;
 mod person;
 mod source;
-
-use xml::Element;
+mod utils;
 
 pub use author::Author;
 pub use category::Category;
@@ -87,50 +86,6 @@ pub use source::Source;
 
 
 const NS: &'static str = "http://www.w3.org/2005/Atom";
-
-
-trait ElementUtils {
-    fn tag_with_text(&mut self, child_name: &'static str, child_body: &str);
-    fn tag_with_optional_text(&mut self, child_name: &'static str, child_body: &Option<String>);
-    fn attribute_with_text(&mut self, attribute_name: &'static str, attribute_value: &str);
-    fn attribute_with_optional_text(&mut self, attribute_name: &'static str, attribute_value: &Option<String>);
-}
-
-
-impl ElementUtils for Element {
-    fn tag_with_text(&mut self, child_name: &'static str, child_body: &str) {
-        self.tag(elem_with_text(child_name, child_body));
-    }
-
-    fn tag_with_optional_text(&mut self, child_name: &'static str, child_body: &Option<String>) {
-        if let Some(ref c) = *child_body {
-            self.tag_with_text(child_name, c);
-        }
-    }
-
-    fn attribute_with_text(&mut self, attribute_name: &'static str, attribute_value: &str) {
-        self.set_attribute(attribute_name.to_string(), None, attribute_value.to_string());
-    }
-
-    fn attribute_with_optional_text(&mut self, attribute_name: &'static str, attribute_value: &Option<String>) {
-        if let Some(ref v) = *attribute_value {
-            self.attribute_with_text(attribute_name, v);
-        }
-    }
-}
-
-
-fn elem_with_text(tag_name: &'static str, chars: &str) -> Element {
-    let mut elem = Element::new(tag_name.to_string(), Some(NS.to_string()), vec![]);
-    elem.text(chars.to_string());
-    elem
-}
-
-
-trait ViaXml where Self: Sized {
-    fn to_xml(&self) -> Element;
-    fn from_xml(elem: Element) -> Result<Self, &'static str>;
-}
 
 
 #[cfg(test)]
